Don't autouse and guard directory removal

+++ tests/conftest.py
@@ -33,7 +33,7 @@ def _create_project(pname=[default_profile_name()]):
     dodocs.main(create_command)
 
 
-@pytest.yield_fixture(scope='session', autouse=True)
+@pytest.yield_fixture(scope='session')
 def homedir(dodocs_homedir):
     """Set the environment variable for the ``dodocs`` home directory and set up
     all the necessary things for running the tests. At the end of the test
@@ -46,7 +46,8 @@ def homedir(dodocs_homedir):
     # create a new project
     _create_project()
     yield dodocs_homedir
-    shutil.rmtree(str(dodocs_homedir))
+    if dodocs_homedir.exists():
+        shutil.rmtree(str(dodocs_homedir))
     del os.environ['DODOCSHOME']
 
 
@@ -54,7 +55,7 @@ def homedir(dodocs_homedir):
 def tmp_homedir(monkeypatch, tmpdir):
     """Override ``homedir`` fixture temporarily unsetting the ``DODOCSHOME``
     environment variable and expanding user directory to a temporary one"""
-    monkeypatch.delitem(os.environ, "DODOCSHOME")
+    monkeypatch.delitem(os.environ, "DODOCSHOME", raising=False)
 
     _tmpdir = Path(str(tmpdir))
